backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-cloud-firestore==2.14.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
apscheduler==3.10.4
python-dotenv==1.0.0
